knowledge-reindexing  - print 추가  - 시작시에만  - 기타 todo 처리  - 주요 type만
 Truncate and padding
  - 최근 대화 순으로 짜르게
  - fix

 BERT-large 용 hidden_size

+++ dataModel.py
@@ -32,8 +32,8 @@ class DialogDataset(Dataset):
 
     def __getitem__(self, idx):
         data = self.train_sample[idx]
-        dialog_token = data['dialog_token']
-        dialog_mask = data['dialog_mask']
+        dialog_token = torch.LongTensor(data['dialog_token'])
+        dialog_mask = torch.LongTensor(data['dialog_mask'])
         target_knowledge = data['target_knowledge']
         response = data['response']
         goal_type = data['goal_type']
